@@ -1,5 +1,5 @@
 module github.com/fushiliang321/jsonrpc
 
-go 1.19
+go 1.21
 
-require github.com/goinggo/mapstructure v0.0.0-20140717182941-194205d9b4a9 // indirect
+require github.com/goinggo/mapstructure v0.0.0-20140717182941-194205d9b4a9
