@@ -2,6 +2,6 @@ package common
 
 import "log"
 
-func Debug(msg interface{}) {
+func Debug(msg any) {
 	log.Println(msg)
 }
