@@ -17,7 +17,7 @@ type Result2 struct {
 
 func (i *IntRpc) Add(params *Params, result *Result) error {
 	a := params.A + params.B
-	*result = interface{}(a).(Result)
+	*result = any(a).(Result)
 	return nil
 }
 
